trying to load images with AlamoImage
