+++ Podfile
@@ -7,5 +7,6 @@ target 'VirtualTourist' do
 
   # Pods for VirtualTourist
   pod 'Alamofire', '~> 3.2.1'
+  pod 'AlamofireImage', '~> 2.0'
 
 end
