+++ VirtualTourist/CollectionView/PictureCell.swift
@@ -1,0 +1,14 @@
+//
+//  PictureCell.swift
+//  VirtualTourist
+//
+//  Created by Max Saienko on 7/7/16.
+//  Copyright © 2016 Max Saienko. All rights reserved.
+//
+
+import UIKit
+
+class PictureCell: UICollectionViewCell {
+    
+    @IBOutlet var picture: UIImageView!
+}
